requirements.txt
asgiref==3.4.1
autopep8==1.6.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.7
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
psycopg2-binary==2.9.1
pycodestyle==2.8.0
python-decouple==3.5
pytz==2021.3
requests==2.26.0
soupsieve==2.3
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
whitenoise==5.3.0
